+++ docs/src/pages/charts/TreemapPage.js
@@ -7,8 +7,6 @@ import PropTable from "../../components/PropTable"
 import LiveExample from "../../components/LiveExample"
 import CodeBlock from "../../components/CodeBlock"
 import PageLayout from "../../components/PageLayout"
-import StreamingToggle from "../../components/StreamingToggle"
-import StreamingDemo from "../../components/StreamingDemo"
 import { Link } from "react-router-dom"
 
 // ---------------------------------------------------------------------------
@@ -72,40 +70,6 @@ const treemapProps = [
 ]
 
 // ---------------------------------------------------------------------------
-// Streaming demo
-// ---------------------------------------------------------------------------
-
-const streamingTreemapCode = `import { Treemap } from "semiotic"
-
-// Update data prop to trigger re-render
-<Treemap data={currentData} childrenAccessor="children" valueAccessor="value" />`
-
-function StreamingTreemapDemo({ width }) {
-  const makeTree = () => ({
-    name: "root",
-    children: [
-      { name: "A", value: Math.round(Math.random() * 200 + 50) },
-      { name: "B", value: Math.round(Math.random() * 200 + 50) },
-      { name: "C", children: [
-        { name: "C1", value: Math.round(Math.random() * 100 + 20) },
-        { name: "C2", value: Math.round(Math.random() * 100 + 20) },
-      ]},
-      { name: "D", value: Math.round(Math.random() * 200 + 50) },
-    ]
-  })
-  const [tree, setTree] = React.useState(makeTree)
-
-  return (
-    <div>
-      <div style={{ marginBottom: 8 }}>
-        <button className="demo-button" onClick={() => setTree(makeTree())}>Randomize Values</button>
-      </div>
-      <Treemap data={tree} width={width} height={400} childrenAccessor="children" valueAccessor="value" nodeIdAccessor="name" colorByDepth />
-    </div>
-  )
-}
-
-// ---------------------------------------------------------------------------
 // Component
 // ---------------------------------------------------------------------------
 
@@ -154,19 +118,17 @@ export default function TreemapPage() {
         each cell proportionally.
       </p>
 
-      <StreamingToggle
-        staticContent={
-          <LiveExample
-            frameProps={{
-              data: sampleData,
-              nodeIdAccessor: "name",
-              valueAccessor: "value",
-              colorBy: "name",
-            }}
-            type={Treemap}
-            startHidden={false}
-            overrideProps={{
-              data: `{
+      <LiveExample
+        frameProps={{
+          data: sampleData,
+          nodeIdAccessor: "name",
+          valueAccessor: "value",
+          colorBy: "name",
+        }}
+        type={Treemap}
+        startHidden={false}
+        overrideProps={{
+          data: `{
   name: "Budget",
   children: [
     { name: "Engineering", children: [
@@ -186,17 +148,9 @@ export default function TreemapPage() {
     { name: "HR", value: 120 }
   ]
 }`,
-              colorBy: '"name"',
-            }}
-            hiddenProps={{}}
-          />
-        }
-        streamingContent={
-          <StreamingDemo
-            renderChart={(w) => <StreamingTreemapDemo width={w} />}
-            code={streamingTreemapCode}
-          />
-        }
+          colorBy: '"name"',
+        }}
+        hiddenProps={{}}
       />
 
       {/* ----------------------------------------------------------------- */}